mpi pp: fix type conversion warning

+++ libs/full/parcelport_mpi/include/hpx/parcelport_mpi/header.hpp
@@ -92,9 +92,9 @@ namespace hpx::parcelset::policies::mpi {
             size_t size = buffer.data_.size();
             size_t numbytes = buffer.data_size_;
             HPX_ASSERT(
-                buffer.num_chunks_.first <= (std::numeric_limits<int>::max)());
+                buffer.num_chunks_.first <= static_cast<unsigned>((std::numeric_limits<int>::max)()));
             HPX_ASSERT(
-                buffer.num_chunks_.second <= (std::numeric_limits<int>::max)());
+                buffer.num_chunks_.second <= static_cast<unsigned>((std::numeric_limits<int>::max)()));
             int num_zero_copy_chunks = buffer.num_chunks_.first;
             int num_non_zero_copy_chunks = buffer.num_chunks_.second;
 
@@ -121,7 +121,7 @@ namespace hpx::parcelset::policies::mpi {
                 size_t tchunk_size = buffer.transmission_chunks_.size() *
                     sizeof(typename parcel_buffer<buffer_type,
                         ChunkType>::transmission_chunk_type);
-                HPX_ASSERT(tchunk_size <= (std::numeric_limits<int>::max)());
+                HPX_ASSERT(tchunk_size <= static_cast<size_t>((std::numeric_limits<int>::max)()));
                 p_format_->numbytes_tchunk = tchunk_size;
                 if (tchunk_size <= max_header_size - current_header_size)
                 {
